Segregate data-uri tests
